requirements.txt
altair
numpy
pandas
pydeck
streamlit
azure-iot-device